+++ test/index.test.js
@@ -1,0 +1,8 @@
+/*global afterEach */
+'use strict';
+var fs = require('fs-extra');
+
+// Remove the `.tmp/` directory after each test.
+afterEach(function () {
+  fs.removeSync('.tmp/');
+});