+++ test/lib/Zip.test.js
@@ -1,14 +1,15 @@
 /*global describe, it */
+'use strict';
 var expect = require('chai').expect;
-var Zip = require('../lib');
+var Zip    = require('../../lib');
 
 describe('Class: `Zip`', function () {
-  
-  it('should be instanciable', function () {
+
+  it('should be a class', function () {
     var zip = new Zip();
     expect(zip).to.be.an.instanceof(Zip);
   });
-  
+
   it('should respond to 7-Zip commands as methods', function () {
     var zip = new Zip();
     expect(zip).to.respondTo('add');
@@ -19,5 +20,5 @@ describe('Class: `Zip`', function () {
     expect(zip).to.respondTo('test');
     expect(zip).to.respondTo('update');
   });
-  
+
 });